Halo Backend/requirements.txt
# Flask Framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Machine Learning
numpy>=1.26.0
pandas>=2.0.3
scikit-learn>=1.3.0
xgboost==1.7.6
lightgbm==4.0.0
joblib==1.3.2

# Deep Learning
tensorflow>=2.16.1


# NLP
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0

#  Database support
psycopg2-binary==2.9.7
SQLAlchemy==2.0.20

# Monitoring
sentry-sdk==1.31.0

# Development
pytest==7.4.2
pytest-cov==4.1.0
